Update GQL schema organization pattern, have each piece export as a complete schema and then use schema merging in the index file to put them together. Possible downside in that we can't use the extends keyword this way, but otherwise it's cleaner and should give clearer error messages than using makeExecutableSchema to put together types and resolvers en masse
